refactor: infrastructure 와 서비스 계층 간의 DTO 생성 및 관련 코드 수정
