+++ src/main/java/teamssavice/ssavice/company/service/client/BusinessVerificationClient.java
@@ -7,5 +7,5 @@ public interface BusinessVerificationClient {
 
 
     // 사업자 등록 정보를 입력받아 진위 여부를 확인
-    CompanyModel.Validate validate(CompanyCommand.Validate command);
+    BusinessVerifyResponse validate(BusinessVerifyRequest request);
 }
