+++ src/main/java/teamssavice/ssavice/company/service/client/BusinessVerifyRequest.java
@@ -1,0 +1,10 @@
+package teamssavice.ssavice.company.service.client;
+
+import lombok.Builder;
+
+@Builder
+public record BusinessVerifyRequest(
+        String businessNumber,
+        String startDate,
+        String name
+) {}
