+++ src/main/java/teamssavice/ssavice/company/service/client/BusinessVerifyResponse.java
@@ -1,0 +1,9 @@
+package teamssavice.ssavice.company.service.client;
+
+import lombok.Builder;
+
+@Builder
+public record BusinessVerifyResponse(
+        boolean isValid
+) {
+}
